refactor(notification): remove ReqType column WISH-225
